+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [Unreleased](https://github.com/kabamgamer/dd2-strategy-hub/tree/HEAD)
+
+* Added search function to all select boxes (defenses, shards, mods)
+
+[Full Changelog](https://github.com/kabamgamer/dd2-strategy-hub/compare/v1.2.1...HEAD)
+
+
+
 ## [v1.2.1](https://github.com/kabamgamer/dd2-strategy-hub/tree/v1.2.1) (2023-09-07)
 
 * Fixed an issue in motherly instinct calculations